Dead-key fix
Previously I thought I had fixed this, the system was posting the
correct messages, however we weren't receiving the correct data.  This
fix ensures that the key press after the accent dead-key in this tested
case '^' followed by a shift key press to capitalize the character, is
receiving the correct values to display in our application.

Also refactored some redundancy