Examples adapted to .net core 3.1 - Debugging in VS is possible now
The ConsoleHook projects have thrown an Invalid Operation exception on Console.ReadKey.  This is solved by starting the 'dotnet' app with the 'consoleHook.dll' as parameter.
The TextBox control in .net core Forms have a different behaviour on handling the '\n' code. So an Environment.NewLine was added afeter each log entry.